Encoder denoiser performance improvement
The denoiser function was modified to reduce the computational
complexity.

1. The denoiser c function modification:
The original code computed the filter_coefficients for different pixel
diff between current raw and last denoised raw, and saved them in
lookup tables. For each pixel c, find
    filter_coefficient[c] = LUT[abs_diff[c]];
and then applying tap coefficients for each pixel.

In order to simplify the process, we calculated the pixel adjustments
vs. pixel diff value ahead of time.
    adjustment = filtered_value - current_raw
               = (filter_coefficient * diff + 128) >> 8

The adjustment vs. diff curve becomes flat very quick when diff increases.
This allowed us to use only several levels to get a close approximation
of the curve. Following the denoiser algorithm, the adjustments are
further modified according to how big the motion magnitude is.

2. The sse2 function was updated.

This change improved the encoder performance by 7% ~ 10% with the
denoiser on.

Change-Id: I93a4308963b8e80c7307f96ffa8b8c667425bf50