New Motion Reference
Alternative strategy for finding a list of candidate motion
vectors to use as reference values in mv coding and as
nearest and near.

Still not implemented in decoder and only 4 of the list of
candidates are currently considered by Yaowu's search code.

Selection code for calculating an optimal vector that
minimizes residual implemented but not used at the moment.
Likewise explicit coding of an optimal reference vector
has not yet been implemented.

Not implemented yet in the decoder but initial tests suggest
an average gain of about 0.5% on derf and gains on almost all
clips. Tests pending forthe other test sets.

Change-Id: I871d7cba1d1c379c4bad9bcccce1fb19c46b8247