Adds feature for companded MV encoding
The high-precision (1/8) pel bit is turned off if the reference
MV is larger than a threshold. The motivation for this patch is
the intuition that if motion is likely large (as indicated by
the reference), there is likley to be more motion blur, and as
a result 1/8 pel precision would be wasteful both in bitrate
as well as computationally.

The feature is incorporated as part of the newmventropy experiment.
There is a modest RD improvement with the patch. Overall the
results with the newmventropy experiment and the threshold being
16 integer pels are below.

derf: +0.279%
std-hd: +0.617%
hd: +1.299%
yt: +0.822%

Further experiments with the threshold, and other variations including
going down to half-pel precision for larger reference MVs are in
progress.

Change-Id: I4ed14600df3c457944e6541ed407cb6e91fe428b