Unit Test For Error Resilient Mode
This unit test compares the difference in quality with
error resilience enabled and disabled. The test runs
for all of the one-pass encoding modes.

The test ensures that the effect of turning on error
resilience makes less than a 10% difference in PSNR.

Further cases shoule be added to do a more comprehensive
test.

Change-Id: Iadc6025ebb60468b85f6f72e7dc23991dec31155