Increase to 8 reference slots
Adds the ability to update up to 8 reference frames on each frame. Any
of the 8 may be selected as one of the 3 active references for each
frame.

This code still has some special cases where slot 0 is treated specially
by the decoder (assumed to be the last frame) for propagating
corruption state for error concealment, though I don't see any references
to it that would be normative. Also, the vpxenc tool makes an assumption
that it contains the last frame for doing the --test-decode check.

Also, updates to slot 2 signal a mode context selection (was
refresh_alt_ref_frame.) It's an open item to determine exactly what this
context switching is supposed to do and make it work for the case of
increased references.

Change-Id: Ie2b910b761ff6fe6ac00f600089005b442936c80