Update patch set 2
Patch Set 2:

I don't think the stats are intended to be parseable by anything else than ourselves. If we ever want to parse it (e.g. for a new unit test), and we insist on reusing this class, we can reintroduce it.

As for a reset, yes, stats_.buf.clear() does the same thing as this patch.

Patch-set: 2