Moved Prediction Filter To be Alternative to Sub-Pel
Previously the prediction filter was optionally applied
before sub-pixel interpolation. To reduce complexity this
patch considers the pred filter as an alternative to
sub-pel interpolation so that the two filtering
operations are applied as alternatives rather than in
combination.

(This patch is a work in progress so I'll -2 it temporarily
until it reaches mergable status!)

Change-Id: Ibc6467d04e4cc80ce452b9db9b9eca4826f610a0