New b-intra mode where direction is contextual
Preliminary patch on a new 4x4 intra mode B_CONTEXT_PRED where the
dominant direction from the context is used to encode. Various decoder
changes were needed to support decoding of B_CONTEXT_PRED in conjunction
with hybrid transforms since the scan order and tokenization depends on
the actual direction of predictioni obtained from the context. Currently
the traditional directional modes are used in conjunction with the
B_CONTEXT_PRED, but it is conceivable that B_CONTEXT_PRED may actually
be used in lieu of the directional modes or at least the less common
oblique directional ones.

Various experiments are underway to find the most efficient combination
of modes to keep, so the patch is still not final. Experimental results
are pending.

Change-Id: I5a7ea80b5218f42a9c0dfb42d3f79a68c7f0cdc2