Add a new token stuffing function vp8_stuff_mb().
This way a caller doesn't need to implement the logic for which (and how
many) tokens to write out to stuff one macroblock worth of EOBs. Make
the actual function implementations static, since they are now only used
in tokenize.c, and do some minor stylistic changes so it follows the
style guide a little more closely.

Change-Id: Ia448e251d19a4e3182eddeb9edd034bd7dc16fa3