Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Which usage of this macro are you seeing the problem with? Is it possible that the extra variable is throwing off the alignment of some other data later in the stack, and that the array this macro is declaring has the correct alignment in either case?

Generally prefer to fix the generic case than add new platform specific code, as this issue may exist on other platforms, if there's something wrong with this macro or how we're using it.

Patch-set: 1
Reviewer: Gerrit User 1000856 <1000856@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1
Label: Verified=0