Update patch set 6
Patch Set 6:

Ok, I hadn't seen that convention. I'm happy to add these consts to the declaration if they're semantically useful.  Probably wouldn't on a public header, but if it's useful to those who know the convention, then that's fine. I just thought we were on a "const ALL the things!!" kick.

Patch-set: 6