Allow for favoring LAST over ALTR/GF prediction.
When LAST and ALTR/GF are available for prediction, allow for the
option of suppressing the prediction off the ALTR/GF under certain conditions.

The current patch allows for suppression under conditions based on delta Q between LAST and ALTR/GF,
and the distance to last update for alt/ref.

The suppression only applies to the case where ALTR/GF was last updated by key frame.

Change-Id: I6509b1d7518bda2e7484cb86ce220b6f77bf1f76