Packing Altref along with succeeding frame and length encoding frames
The altref frame is packed along with the next P frame. So that
outside of the codec there are now only two types of frames P and I.
Also, now it is one frame in and one frame out with respect to the
codec. Apart from that, all the frames are length encoded. There are
two categories of frames and each of them will look as follows:

  - Packed frames (an altref along with the succeeding p frame)
    - control byte | length | altref | control byte | length | frame
    data
  - Unpacked frames (all frames other than the above)
    - control byte | length | frame data

Change-Id: If1eabf5c473f7d46b3f2d026bd30c803588c5330