Reworking the compound intra-intra experiment
Incorporates the intra-intra compound prediction in B_PRED modes of
both key and inter frames. A MB level flag indicates if any of the
16 4x4 blocks use a second intra predictor or not, and if the flag
is 1, an additional bit is sent with each 4x4 b-mode to signal
existence of a second b-mode.
Adds probability updates for the Mb and block level flags, and also
handles costing for rd optimization appropriately.

Change-Id: I5213bb07aa029d678ac99b4f128a13b1048693fb