Reworking the compound intra-intra experiment
Incorporates the intra-intra compound prediction in B_PRED modes of
both key and inter frames. A MB level flag indicates if any of the
16 4x4 blocks use a second intra predictor or not; and if this flag
is 1, an additional bit is sent with each 4x4 b-mode to signal
existence of a second b-mode. Besides a frame level flag indicates
if compound intra is to be used at all or not.
Adds probability updates for the Mb and block level flags, and also
handles costing for rd optimization appropriately.
The algorithm for combining two intra predictors is also modified to
handle pixels differently where the two predictors do not agree, by
using the TM_PRED predictor as a cue.
Currently only B_PRED modes are handled, but the experiment will later
include 16x16 modes as well.

Change-Id: I5213bb07aa029d678ac99b4f128a13b1048693fb