fixed a mismatch between 32 and 64 bit builds
All right, we do not necessarily ever merge this commit, but I like
to get a record of this. After a lot of back and forth, scott and I
were able to narrow down a 32bit/64bit build mismatch issue to the
function of 16x16 idct which uses double/floating point operations.
Suggestedn by James Zern that the issue may be caused by different
precision depednign on if the double/floating point operation are
executed through 80bit x87 fpu or 64bit SSE uint.

And of course, the 16x16 idct function itself will be integerized
soon, and the double/floating point should not be an real issue
regardless.

Change-Id: I37e17298aed7dee59d074efbc6b802b3529e0047