Compound inter-intra experiment
A patch on compound inter-intra prediction.
In compound inter-intra prediction, a new predictor is
obtained by combining the best inter predictor with the
best 16x16 intra predictor, in a manner that the weight
varies with distance from the top/left boundary.

The patch can be reviewed now. Results are pending.

Change-Id: I7951f1ed0d6eb31ca32ac24d120f1585bcd8d79b