Compound inter-intra experiment
A patch on compound inter-intra prediction.
In compound inter-intra prediction, a new predictor for
16x16 inter coded MBs are obtained by combining the best
inter predictor with the best 16x16 intra predictor, in
a manner that the weight varies with distance from the
top/left boundary.

The patch can be reviewed now, even though some experiments
are pending.

Results so far:

derf +0.28%
yt +0.32%
std-hd +0.35%
hd +0.42%

Change-Id: I7951f1ed0d6eb31ca32ac24d120f1585bcd8d79b