Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

(2 inline comments)

This comment doesn't block anything, but I wanted to make it anyway -- does this mean the meaning of xd->mb_to_right_edge etc changes now depending on context? I always thought of it as the distance to the border of the image, but now it's different depending on the mode of the current block. I wonder if it wouldn't be better to keep it with the current, consistent meaning, and change the thresholds in the sb-specific contexts where the difference matters (eg, xd->mb_to_right_edge <= (16<<3) or equivalent).

I didn't trace through to see if there were any implications on clamping/entropy/etc, but this is an area that's horrendously complicated/buggy/odd in VP8, and should get careful evaluation at some point.

-1 only for the mb_col bug, just mho on the above.

Patch-set: 3
Reviewer: Gerrit User 1000856 <1000856@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1
Label: Verified=0