Update patch set 4
Patch Set 4:

So, re: your earlier comment (@jkoleszar), the original SB code actually did the wrong thing and changed the meaning of mb_to_right/bottom_edge. The result of that was that left/top MV clamping didn't work and would lead to "wrong" (well, unideal) edge extension, it would basically use the 16 pixels within the visible area and 16 pixels of the edge for the 32x32 predictor, which is obviously wrong.

This commit tries to fix that, and make sure that mb_to_*_edge has the same meaning between sb and mb. Whether that meaning is correct (in terms of problems that we had in vp8) is probably something we should investigate, as you said.

Patch-set: 4
Reviewer: Gerrit User 1000885 <1000885@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Verified=0