Update patch set 1
Patch Set 1:

There is no top-level cpu-features.h. It's under machine/ (or asm/ for mips). In a normal NDK build, this path is added automatically. However, when building from the gyp the file is not found. In the past they added machine/ to the include path, but that creates issues with setjmp.h (which is in both).

It's guarded by CHROMIUM_BUILD to make it explicit that it is a hack. I can add documentation to the source file if that makes things clearer.

Patch-set: 1
Reviewer: Gerrit User 1117029 <1117029@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=0