New previous coef context experiment
Adds an experiment to derive the previous context of a coefficient not
just from the previous coefficient in the scan order but from a combination
of several neighboring coefficients previously encountered in scan order.i
A precomputed table of neighbors for each location for each scan type and
block size is used. Currently 5 neighbors are used.
Trellis quantization is still the old one and needs update.

Results are pending.

Change-Id: Ie708b54d8e1898af742846ce2d1e2b0d89fd4ad5