New previous coef context experiment
Adds an experiment to derive the previous context of a coefficient not
just from the previous coefficient in the scan order but from a combination
of several neighboring coefficients previously encountered in scan order.
A precomputed table of neighbors for each location for each scan type and
block size is used. Currently 5 neighbors are used.
Trellis quantization is still the old one and has not been updated yet.

Results are about 0.2% positive using a strategy where the max coef magnitude
from the 5 neigbors is used to derive the context.

Change-Id: Ie708b54d8e1898af742846ce2d1e2b0d89fd4ad5