Update patch set 11
Patch Set 11:

Re: the remainder of the style warnings. vp9_entropy.c is almost exclusively the zigzag table. I'd like to keep it as-is so it's in a 32x32 format. The rest of the warnings are either hard to prevent because of #if macros, or are fixed in my follow-up patch that adds vp9_coeff_stats/accum/probs/count.

Patch-set: 11
Label: Code-Review=0
Label: Verified=0