Add multisymbol range coder experiment
Adds a new 'multisymbol' experiment to allow SIMD-friendly symbol decoding.
This patch replaces the existing boolhuff/dboolhuff coders using the new
range coder, but exposes it with the existing API so that minimal additional
code changes are required.

For further background, see:
  http://tools.ietf.org/html/draft-terriberry-codingtools-00
  https://people.xiph.org/~tterribe/daala/ec_test0/ec_test.c

This decoder is ~1-5% faster in a microbenchmark against the existing bool
decoder, though it's slower in the actual VP9 decoder. The primary intent
at this point is to get complexity gains and non-boolean contexts, so not
trying to microoptimize decode_bool at this point.

Change-Id: I196b2342802690e0e97b12ae8d221cec0a2f66f2