Reset index so mode evaluation loop will move on
There are circumstances where the evaluation of coding
modes can enter an infinite loop.

When we hit a "continue" in the loop we should reset
switchable_filter_index so tha the loop counter moves
onto the next mode.

This patch resets the index in one circumstance, but
it may also have to be reset in the other cases where
a continue statement is found in the loop.

Change-Id: I39256d1c9a73264d707109115d7e309912c50616