Replace 6-tap filter with 8-tap Low-Pass filter
This patch removes the six-tap filter and replaces it with an
eight-tap low-pass filter.

If SWITCHABLE is enabled as the default interpolation filter
and VP9_SWITCHABLE_FILTERS is set to 3, then 3 alternative
interpolation filters are tested in the inter-mode selection
code. This patch adds a low-pass variant to the standard
interpolating and high-pass variants.

This implements the first stage of the new filtered prediction
strategy. The 8-tap low-pass filter implements the strategy
for non-pixel aligned motion vectors. Later patches will add
the option to filter at pel-aligned positions and also for
intra-coded predictors.

A further future patch will remove a layer from the filtering
logic that converts from a fractional pel motion vector
component to a pointer to the required filter taps. In this way
the three sets of functions in vp9_filter.c can be combined.

NOTE: The PPC build has not been modified to reflect the
content of this patch since it is no longer functional.

Change-Id: Ifbba9934932727a7ac258b816cdeb500362fc246