Supporting mutiple coef context models per frame
Allows multiple coef context models to be tracked and updated
per frame. The motivation is better performance for high-definition
content as well as parallel decodability. The context zones are
aligned along superblock rows and are all of the same size (except
the last which can be smaller).

A 2 bit index is specified along wih every key frame to indicate
how many adaptation zones (0-3) are to be used until the next key
frame.  If this field is set as 0, backward adaptation is turned off.

Another bit specifies if forward updates are to be turned off.

Currently only coeff contexts are updated. But later mode/mvs can
also be folded into the framework.

Change-Id: Iecc8089940e62d28c487c5a962101cdbe4339cf2