Supporting mutiple coef context models per frame
Allows multiple coef context models to be tracked and updated
per frame. The motivation is better performance for high-definition
content. The context zones are aligned along superblock rows and
are all of the same size (except the last which can be smaller).

A 3 bit index is specified along wih every key frame to indicate
how many adaptation zones are to be used until the next key frame.
If this field is set as zero, backward adaptation is turned off
(which may be helpful for parallel decoding).

Change-Id: Iecc8089940e62d28c487c5a962101cdbe4339cf2