Update patch set 3
Patch Set 3:

I like the idea of being able to select which type of updates we support. From this, is it possible to gather how much gains we're actually getting in practice from backwards and forward updates (and would it be possible to add a commandline argument to vpxenc to be able to selective disable the backward updates)?

Patch-set: 3
Reviewer: Gerrit User 1000885 <1000885@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0